Mkexy build require paths so exerb include them properly
mkexy is not aware that Ruby 1.9 expand $LOADED_FEATURES elements and
includes the full path to them.

require_name inside Utility2 tries to find a match of the right $LOAD_PATH
and remove it from the final require name.

It also removes the RUBY_PLATFROM from it so shared libraries are found.